Extracting AppData record from type declaration
